New type of StumpWM builds: build with preloaded (built-in) modules.
You can specify modules you want to be preloaded in stumpwm in
builtin-modules.lisp-expr file in this directory which contains a list of
module names, in example, ("foo" "bar" "cool-module"). These modules will be
initialized when you launch stumpwm, no call to load-module is required in your
.stumpwmrc file.

Bug: Some modules (like wallpaper module) can not be built-in in this way,
because they require late initialization and built-in modules all are
initialized very early on StumpWM startup.